requirements/base.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # For PostgreSQL

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Utils
email-validator==2.1.0.post1
pydantic-settings==2.2.1
python-slugify==8.0.1
